backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyconfig==0.13.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
av==11.0.0
backoff==2.2.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.12.1
blis==0.7.11
boto3==1.34.1
botocore==1.34.1
build==1.0.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cohere==4.47
colorama==0.4.6
colorlog==6.8.0
comm==0.2.2
confection==0.1.4
contourpy==1.2.0
cookiecutter==2.5.0
courlan==0.9.5
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
distro==1.8.0
dnspython==2.6.1
docker-pycreds==0.4.0
dynaconf==3.2.4
easyocr==1.7.1
emoji==2.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.105.0
fastavro==1.9.4
filelock==3.13.1
filetype==1.2.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
htmldate==1.6.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
imageio==2.33.1
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jusText==3.0.0
kedro==0.18.14
kiwisolver==1.4.5
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.70
lazy_loader==0.3
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.2
omegaconf==2.3.0
openai==1.12.0
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
parse==1.20.0
parso==0.8.3
pathspec==0.12.1
pdf2image==1.16.3
pdfminer==20191125
pexpect==4.9.0
Pillow==10.1.0
pip-tools==7.3.0
platformdirs==4.1.0
pluggy==1.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.6.1
pydantic_core==2.16.2
pyfiglet==1.0.2
Pygments==2.17.2
pymongo==4.8.0
pyparsing==3.1.2
pypdf==3.17.2
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pytest==7.4.4
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-slugify==8.0.1
python-telegram-bot==20.8
pytoolconfig==1.2.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.5.2
redis==5.0.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rope==1.11.0
s3transfer==0.9.0
schedule==1.2.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
semantic-router==0.0.20
sentry-sdk==1.39.2
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
slack-bolt==1.18.1
slack-sdk==3.26.1
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spotipy==2.23.0
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2023.12.9
tld==0.13
toml==0.10.2
tomli==2.0.1
toposort==1.10
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
trafilatura==1.6.4
traitlets==5.14.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.5
unstructured-client==0.15.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
Wand==0.6.13
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.12
weasel==0.3.4
weaviate-client==3.25.3
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
